requirements.txt
docutils==0.12
gitdb==0.6.4
GitPython==2.0.5
markdown2==2.3.1
Pygments==2.1.3
smmap==0.9.0
tornado==4.3
